PDDA-474 Moving getCourtSites() out of CourtRoomService into AbstractService
