+++ src/main/java/uk/gov/hmcts/pdm/publicdisplay/manager/service/CourtRoomService.java
@@ -8,10 +8,8 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.pdm.business.entities.xhbcourt.XhbCourtDao;
 import uk.gov.hmcts.pdm.business.entities.xhbcourtroom.XhbCourtRoomDao;
-import uk.gov.hmcts.pdm.business.entities.xhbcourtsite.XhbCourtSiteDao;
 import uk.gov.hmcts.pdm.publicdisplay.manager.dto.CourtDto;
 import uk.gov.hmcts.pdm.publicdisplay.manager.dto.CourtRoomDto;
-import uk.gov.hmcts.pdm.publicdisplay.manager.dto.XhibitCourtSiteDto;
 import uk.gov.hmcts.pdm.publicdisplay.manager.security.UserRole;
 import uk.gov.hmcts.pdm.publicdisplay.manager.service.api.ICourtRoomService;
 import uk.gov.hmcts.pdm.publicdisplay.manager.web.courtroom.CourtRoomAmendCommand;
@@ -25,18 +23,12 @@ import java.util.Optional;
 
 @Component
 @Transactional(readOnly = true, propagation = Propagation.REQUIRED)
-public class CourtRoomService extends CourtRoomServiceFinder implements ICourtRoomService {
+public class CourtRoomService extends CourtRoomServiceCreator implements ICourtRoomService {
 
     /**
      * Set up our logger.
      */
     private static final Logger LOGGER = LoggerFactory.getLogger(CourtRoomService.class);
-    private static final String METHOD = "Method ";
-    private static final String THREE_PARAMS = "{}{}{}";
-    private static final String FOUR_PARAMS = "{}{}{}{}";
-    private static final String STARTS = " - starts";
-    private static final String ENDS = " - ends";
-    private static final String YES = "Y";
     private static final String EMPTY_STRING = "";
     private static final String REGEXP_NUMBERS = "[^0-9]";
     private static final Integer ROOM_NO_LIMIT = 99;
@@ -74,52 +66,6 @@ public class CourtRoomService extends CourtRoomServiceFinder implements ICourtRo
             // Sort by name
             Collections.sort(resultList, (obj1, obj2) -> String.CASE_INSENSITIVE_ORDER
                 .compare(obj1.getCourtName(), obj2.getCourtName()));
-        }
-        LOGGER.info(THREE_PARAMS, METHOD, methodName, ENDS);
-        return resultList;
-    }
-
-    /**
-     * Gets the court sites.
-     *
-     * @return the court sites
-     */
-    @Override
-    public List<XhibitCourtSiteDto> getCourtSites(Integer courtId) {
-        final String methodName = "getCourtSites";
-        LOGGER.info(THREE_PARAMS, METHOD, methodName, STARTS);
-        final List<XhibitCourtSiteDto> resultList = new ArrayList<>();
-        final List<XhbCourtSiteDao> xhibitCourtSiteList;
-
-        if (courtId == null) {
-            xhibitCourtSiteList = getXhbCourtSiteRepository().findAll();
-        } else {
-            xhibitCourtSiteList = getXhbCourtSiteRepository().findByCourtId(courtId);
-        }
-        LOGGER.debug(FOUR_PARAMS, METHOD, methodName, " - Court sites returned : ",
-            xhibitCourtSiteList.size());
-
-        if (!xhibitCourtSiteList.isEmpty()) {
-            // Transfer each court site to a dto and save in resultList
-            for (XhbCourtSiteDao xhibitCourtSite : xhibitCourtSiteList) {
-                if (YES.equals(xhibitCourtSite.getObsInd())) {
-                    continue;
-                }
-                LOGGER.debug(THREE_PARAMS, METHOD, methodName, " - transferring court site to dto");
-                final XhibitCourtSiteDto dto = createXhibitCourtSiteDto();
-
-                // need the court site details from the main court site in 'xhb_court_site' table
-                dto.setId(xhibitCourtSite.getId().longValue());
-                dto.setCourtSiteName(xhibitCourtSite.getCourtSiteName());
-                dto.setCourtSiteCode(xhibitCourtSite.getCourtSiteCode());
-                dto.setCourtId(xhibitCourtSite.getCourtId());
-                LOGGER.debug("dto id : {}", dto.getId());
-                LOGGER.debug("dto courtSiteName: {}", dto.getCourtSiteName());
-                resultList.add(dto);
-            }
-            // Sort by name
-            Collections.sort(resultList, (obj1, obj2) -> String.CASE_INSENSITIVE_ORDER
-                .compare(obj1.getCourtSiteName(), obj2.getCourtSiteName()));
         }
         LOGGER.info(THREE_PARAMS, METHOD, methodName, ENDS);
         return resultList;
